Move FilteredSelect to wagtail.admin.widgets
