+++ wagtail/admin/widgets/__init__.py
@@ -3,4 +3,5 @@ from wagtail.admin.widgets.button import *  # NOQA
 from wagtail.admin.widgets.button_select import *  # NOQA
 from wagtail.admin.widgets.chooser import *  # NOQA
 from wagtail.admin.widgets.datetime import *  # NOQA
+from wagtail.admin.widgets.filtered_select import *  # NOQA
 from wagtail.admin.widgets.tags import *  # NOQA
